Clean up more elements for include-headers
