+++ xtrack/beam_elements/elements_src/drift.h
@@ -7,48 +7,20 @@
 #define XTRACK_DRIFT_H
 
 #include <headers/track.h>
+#include <beam_elements/elements_src/track_drift.h>
 
 
 GPUFUN
-void Drift_single_particle_expanded(LocalParticle* part, double length){
-    double const rpp    = LocalParticle_get_rpp(part);
-    double const rv0v    = 1./LocalParticle_get_rvv(part);
-    double const xp     = LocalParticle_get_px(part) * rpp;
-    double const yp     = LocalParticle_get_py(part) * rpp;
-    double const dzeta  = 1 - rv0v * ( 1. + ( xp*xp + yp*yp ) / 2. );
+void Drift_track_local_particle(DriftData el, LocalParticle* part0){
 
-    LocalParticle_add_to_x(part, xp * length );
-    LocalParticle_add_to_y(part, yp * length );
-    LocalParticle_add_to_s(part, length);
-    LocalParticle_add_to_zeta(part, length * dzeta );
-}
-
-
-GPUFUN
-void Drift_single_particle_exact(LocalParticle* part, double length){
-    double const px = LocalParticle_get_px(part);
-    double const py = LocalParticle_get_py(part);
-    double const rv0v    = 1./LocalParticle_get_rvv(part);
-    double const one_plus_delta = 1. + LocalParticle_get_delta(part);
-
-    double const one_over_pz = 1./sqrt(one_plus_delta*one_plus_delta
-                                       - px * px - py * py);
-    double const dzeta = 1 - rv0v * one_plus_delta * one_over_pz;
-
-    LocalParticle_add_to_x(part, px * one_over_pz * length);
-    LocalParticle_add_to_y(part, py * one_over_pz * length);
-    LocalParticle_add_to_zeta(part, dzeta * length);
-    LocalParticle_add_to_s(part, length);
-}
-
-
-GPUFUN
-void Drift_single_particle(LocalParticle* part, double length){
-    #ifndef XTRACK_USE_EXACT_DRIFTS
-        Drift_single_particle_expanded(part, length);
-    #else
-        Drift_single_particle_exact(part, length);
+    double length = DriftData_get_length(el);
+    #ifdef XSUITE_BACKTRACK
+        length = -length;
     #endif
+
+    START_PER_PARTICLE_BLOCK(part0, part);
+        Drift_single_particle(part, length);
+    END_PER_PARTICLE_BLOCK;
 }
