+++ xtrack/beam_elements/elements_src/marker.h
@@ -1,0 +1,13 @@
+// copyright ############################### //
+// This file is part of the Xtrack Package.  //
+// Copyright (c) CERN, 2021.                 //
+// ######################################### //
+
+#ifndef XTRACK_MARKER_H
+#define XTRACK_MARKER_H
+
+#include <headers/track.h>
+
+GPUFUN void Marker_track_local_particle(MarkerData el, LocalParticle* part0) {}
+
+#endif  // XTRACK_MARKER_H
