+++ xtrack/beam_elements/elements_src/multipoleedge.h
@@ -7,6 +7,7 @@
 #define XTRACK_MULTIPOLEEDGE_H
 
 #include <headers/track.h>
+#include <beam_elements/elements_src/track_mult_fringe.h>
 
 
 GPUFUN
