+++ xtrack/beam_elements/elements_src/wedge.h
@@ -6,6 +6,7 @@
 #define XTRACK_WEDGE_H
 
 #include <headers/track.h>
+#include <beam_elements/elements_src/track_wedge.h>
 
 
 GPUFUN
