+++ xtrack/beam_elements/elements_src/xrotation.h
@@ -6,6 +6,7 @@
 #define XTRACK_XROTATION_H
 
 #include <headers/track.h>
+#include <beam_elements/elements_src/track_xrotation.h>
 
 
 GPUFUN
