+++ xtrack/beam_elements/slice_elements_thick.py
@@ -179,10 +179,8 @@ class ThickSliceSolenoid(BeamElement):
     _internal_record_class = SynchrotronRadiationRecord
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('headers/constants.h'),
-        _pkg_root.joinpath('headers/synrad_spectrum.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/thick_slice_solenoid.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/thick_slice_solenoid.h'),
+    ]
 
     copy = _slice_copy
 
@@ -211,8 +209,8 @@ class DriftSliceBend(BeamElement):
     _xofields = {'_parent': xo.Ref(Bend), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_bend.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_bend.h'),
+    ]
 
     def to_dict(self, **kwargs):
         dct = BeamElement.to_dict(self, **kwargs)
@@ -246,8 +244,8 @@ class DriftSliceRBend(BeamElement):
     _xofields = {'_parent': xo.Ref(RBend), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_rbend.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_rbend.h'),
+    ]
 
     def to_dict(self, **kwargs):
         dct = BeamElement.to_dict(self, **kwargs)
@@ -281,8 +279,8 @@ class DriftSliceQuadrupole(BeamElement):
     _xofields = {'_parent': xo.Ref(Quadrupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_quadrupole.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_quadrupole.h'),
+    ]
 
     copy = _slice_copy
 
@@ -316,8 +314,8 @@ class DriftSliceSextupole(BeamElement):
     _xofields = {'_parent': xo.Ref(Sextupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_sextupole.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_sextupole.h'),
+    ]
 
     copy = _slice_copy
 
@@ -351,8 +349,8 @@ class DriftSliceOctupole(BeamElement):
     _xofields = {'_parent': xo.Ref(Octupole), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_octupole.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice_octupole.h'),
+    ]
 
     copy = _slice_copy
 
@@ -386,8 +384,8 @@ class DriftSlice(BeamElement):
     _xofields = {'_parent': xo.Ref(Drift), **COMMON_SLICE_XO_FIELDS}
 
     _extra_c_sources = [
-        _pkg_root.joinpath('beam_elements/elements_src/drift.h'),
-        _pkg_root.joinpath('beam_elements/elements_src/drift_slice.h')]
+        _pkg_root.joinpath('beam_elements/elements_src/drift_slice.h'),
+    ]
 
     copy = _slice_copy
